requirements.txt
alabaster==0.7.13
anyio==3.7.0
Babel==2.12.1
bleach==6.0.0
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.1.0
chatterbot-corpus==1.2.0
click==8.1.3
colorama==0.4.6
confection==0.0.4
coverage==6.5.0
coveralls==3.3.1
cymem==2.0.7
DAWG-Python==0.7.2
docopt==0.6.2
docutils==0.18.1
emoji==2.2.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
exceptiongroup==1.1.1
fastapi==0.89.1
filelock==3.9.1
flake8==6.0.0
Flask==2.2.3
h11==0.14.0
idna==3.4
imagesize==1.4.1
importlib-metadata==6.6.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
Jinja2==3.1.2
joblib==1.2.0
keyring==23.13.1
langcodes==3.3.0
language-tool-python==2.7.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mathparse==0.1.2
mccabe==0.7.0
mdurl==0.1.2
more-itertools==9.1.0
murmurhash==1.0.9
nltk==3.8.1
nose==1.3.7
numpy==1.23.5
oauthlib==3.2.2
packaging==23.1
pandas==1.5.3
pathy==0.10.1
Pint==0.22
pkginfo==1.9.6
preshed==3.0.8
prometheus-client==0.17.0
pybind11==2.10.3
pycodestyle==2.10.0
pydantic==1.10.8
pyflakes==3.0.1
Pygments==2.15.1
pymongo==3.13.0
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pymystem3==0.2.0
pyTelegramBotAPI==4.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32-ctypes==0.2.0
PyYAML==5.3.1
readme-renderer==37.3
regex==2023.6.3
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.4.1
scikit-learn==1.0.2
scipy==1.9.3
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
snowballstemmer==2.2.0
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
Sphinx==3.0.4
sphinx-rtd-theme==1.2.1
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.3.24
srsly==2.4.6
starlette==0.22.0
thinc==8.1.10
threadpoolctl==3.1.0
tqdm==4.64.1
twine==4.0.2
twython==3.9.1
typer==0.7.0
typing_extensions==4.6.3
urllib3==1.26.15
uvicorn==0.18.3
wasabi==1.1.1
webencodings==0.5.1
Werkzeug==2.2.3
zipp==3.15.0
